Crate pallet_plasm_staking

Plasm Staking Module

The Plasm staking module manages era, total amounts of rewards and how to distribute.

Modules

parameters
Plasm Staking Parameters

Structs

Bonded
Map from all locked "stash" accounts to the controller account.

CurrentElected
The currently elected validator set keyed by stash account ID.

CurrentEra
The current era index.

CurrentEraStart
The start of the current era.

CurrentEraStartSessionIndex
The session index at which the current era started.

Exposure
A snapshot of the stake backing a single validator in the system.

ForceEra
True if the next session change will be a new era regardless of index.

GenesisConfig

IndividualExposure
The amount of exposure (to slashing) than an individual nominator has.

Ledger
Map from all (unlocked) "controller" accounts to the info regarding the staking.

Module

Nominations
A record of the nominations made by a specific account.

Payee
Where the reward payment should be made. Keyed by stash.

StakedContracts
Nominators for a particular contract that is in action right now.

StakingLedger
The ledger of a (bonded) stash.

StorageVersion
The version of storage for upgrade.

UnlockChunk
Just a Balance/BlockNumber tuple to encode when a chunk of funds will be unlocked.

Validators
Set of next era accounts that can validate blocks.

Enums

Call
Dispatchable calls.

Forcing
Mode of era-forcing.

RawEvent
Events for this module.

RewardDestination
A destination account for payment.

Traits

Trait

Type Definitions

BalanceOf

EraIndex

Event
RawEvent specialized for the configuration Trait

MomentOf

SessionIndex
Simple index type with which we can count sessions.