Struct pallet_plasm_staking::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: Trait + 'static> Module<T>

pub fn bonded<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::AccountId>

Map from all locked "stash" accounts to the controller account.

pub fn ledger<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>>

Map from all (unlocked) "controller" accounts to the info regarding the staking.

pub fn payee<K: EncodeLike<T::AccountId>>(key: K) -> RewardDestination

Where the reward payment should be made. Keyed by stash.

pub fn dapps_nominations<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<Nominations<T::AccountId>>

The map from nominator stash key to the set of stash keys of all contracts to nominate.

NOTE: is private so that we can ensure upgraded before all typical accesses. Direct storage APIs can still bypass this protection.

pub fn staked_contracts<K: EncodeLike<T::AccountId>>(
key: K
) -> Exposure<T::AccountId, BalanceOf<T>>

Nominators for a particular contract that is in action right now.

This is keyed by the contract account id.

pub fn current_elected() -> Vec<T::AccountId>

The currently elected validator set keyed by stash account ID.

pub fn current_era() -> EraIndex

The current era index.

pub fn current_era_start() -> MomentOf<T>

The start of the current era.

pub fn current_era_start_session_index() -> SessionIndex

The session index at which the current era started.

pub fn force_era() -> Forcing

True if the next session change will be a new era regardless of index.

pub fn storage_version() -> u32

The version of storage for upgrade.

pub fn validators() -> Vec<T::AccountId>

Set of next era accounts that can validate blocks.

impl<T: Trait> Module<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn new_session(
ending: SessionIndex,
will_apply_at: SessionIndex
) -> Option<Vec<T::AccountId>>

Session has just ended. Provide the validator set for the next session if it's an era-end, along with the exposure of the prior validator set.

pub fn new_era(
_ending: SessionIndex,
will_apply_at: SessionIndex
) -> Option<Vec<T::AccountId>>

The era has changed - enact new staking set.

NOTE: This always happens immediately before a session change to ensure that new validators get a chance to set their session keys.

pub fn reward_to_validators(
total_payout: BalanceOf<T>,
max_payout: BalanceOf<T>
) -> BalanceOf<T>

pub fn reward_to_operators(
total_payout: BalanceOf<T>,
max_payout: BalanceOf<T>
) -> BalanceOf<T>