Trait pallet_plasm_staking::Trait

pub trait Trait: Trait {
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type ContractFinder: ContractFinder<Self::AccountId, StakingParameters>;
    type BondingDuration: Get<EraIndex>;
    type RewardRemainder: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type Reward: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>;
    type Time: Time;
    type SessionsPerEra: Get<SessionIndex>;
    type Event: From<Event<Self>> + Into<Self::Event>;
}

Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type ContractFinder: ContractFinder<Self::AccountId, StakingParameters>

type BondingDuration: Get<EraIndex>

Number of eras that staked funds must remain bonded for.

type RewardRemainder: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Tokens have been minted and are unused for validator-reward. Maybe, plasm-staking uses ().

type Reward: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>

Handler for the unbalanced increment when rewarding a staker. Maybe, plasm-staking uses ().

type Time: Time

Time used for computing era duration.

type SessionsPerEra: Get<SessionIndex>

Number of sessions per era.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

Loading content...